requirements.txt
requests>=2.28.0
websockets>=10.0
pyserial>=3.5 